moved LoginPage and added stylesheet. Moved some styles out of bulma
