+++ client/src/app.scss
@@ -1,9 +1,10 @@
-$secondary: rgb(56, 0, 17);
+$secondaryColor: rgb(56, 0, 17);
 //orange
-// $secondary: rgb(241, 146, 3);
+// $secondaryColor: rgb(241, 146, 3);
 //blue
-// $secondary: rgb(0, 183, 255);
-$tertiary: lighten(desaturate($secondary, 80%), 70);
+// $secondaryColor: rgb(0, 183, 255);
+$secondary: lighten($secondaryColor,0);
+$tertiary: lighten(desaturate($secondaryColor, 80%), 70);
 $primary: rgba(242, 242, 242, 0.95);
 //smaller size for commenting buttons
 $smaller: 2.5em;
@@ -20,7 +21,7 @@ $family-secondary: "Indie Flower", cursive;
 @import "../../node_modules/bulma/sass/elements/content.sass";
 
 body {
-  background: lighten($secondary, 0);
+  background: desaturate(lighten($secondary, 0),50);
   font-family: "Montserrat", sans-serif;
   height: var(--bodyHeight);
   overflow: hidden;
