+++ client/src/pages/LoginPage/loginPage.scss
@@ -1,0 +1,18 @@
+@import "../../app.scss";
+
+#loginButton, #signupButton {
+    margin-top: 0.5rem;
+    background-color: darken(desaturate($secondary, 0), 0);
+    color: $light;
+  }
+  #loginButton:hover,
+  #loginButton:focus,
+  #signupButton:hover,
+  #signupButton:focus {
+    background-color:
+      lighten(
+        desaturate($secondary, 10),
+        10
+      );
+    color: $light;
+  }
